feat: add os labeling and refactor
