--- .github/labeler.yml
@@ -1,2 +1,0 @@
-bug:
-  - .github/ISSUE_TEMPLATE/bug.yml
